cv-service-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0

opencv-python>=4.10.0
numpy>=2.0
Pillow>=10.0.0

pydantic==2.5.0
python-multipart==0.0.6

scikit-image>=0.22.0